# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NeuroVerse",
    "description": "Multimodal Biosensing and Adaptive Neuroadaptive Interface Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/acquisition/test": {
      "post": {
        "tags": [
          "acquisition"
        ],
        "summary": "Api Acquisition Test",
        "operationId": "api_acquisition_test_api_v1_acquisition_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AcquisitionSelectRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Api Acquisition Test Api V1 Acquisition Test Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/acquisition/select": {
      "post": {
        "tags": [
          "acquisition"
        ],
        "summary": "Api Acquisition Select",
        "operationId": "api_acquisition_select_api_v1_acquisition_select_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcquisitionSelectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Api Acquisition Select Api V1 Acquisition Select Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/acquisition/lsl/streams": {
      "get": {
        "tags": [
          "acquisition"
        ],
        "summary": "Api Lsl Streams",
        "operationId": "api_lsl_streams_api_v1_acquisition_lsl_streams_get",
        "parameters": [
          {
            "name": "stream_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stream Name"
            }
          },
          {
            "name": "stream_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stream Type"
            }
          },
          {
            "name": "source_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          },
          {
            "name": "timeout_seconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1.0,
              "title": "Timeout Seconds"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Api Lsl Streams Api V1 Acquisition Lsl Streams Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/acquisition/lsl/streams/{stream_id}/metadata": {
      "get": {
        "tags": [
          "acquisition"
        ],
        "summary": "Api Lsl Stream Metadata",
        "operationId": "api_lsl_stream_metadata_api_v1_acquisition_lsl_streams__stream_id__metadata_get",
        "parameters": [
          {
            "name": "stream_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Stream Id"
            }
          },
          {
            "name": "timeout_seconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1.0,
              "title": "Timeout Seconds"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Api Lsl Stream Metadata Api V1 Acquisition Lsl Streams  Stream Id  Metadata Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/acquisition/brainflow/test": {
      "post": {
        "tags": [
          "acquisition"
        ],
        "summary": "Api Brainflow Test",
        "operationId": "api_brainflow_test_api_v1_acquisition_brainflow_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrainFlowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Api Brainflow Test Api V1 Acquisition Brainflow Test Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/acquisition/brainflow/select": {
      "post": {
        "tags": [
          "acquisition"
        ],
        "summary": "Api Brainflow Select",
        "operationId": "api_brainflow_select_api_v1_acquisition_brainflow_select_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrainFlowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Api Brainflow Select Api V1 Acquisition Brainflow Select Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/acquisition/brainflow/start": {
      "post": {
        "tags": [
          "acquisition"
        ],
        "summary": "Api Brainflow Start",
        "operationId": "api_brainflow_start_api_v1_acquisition_brainflow_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrainFlowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Api Brainflow Start Api V1 Acquisition Brainflow Start Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/acquisition/lsl/select": {
      "post": {
        "tags": [
          "acquisition"
        ],
        "summary": "Api Lsl Select",
        "operationId": "api_lsl_select_api_v1_acquisition_lsl_select_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LslSelectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Api Lsl Select Api V1 Acquisition Lsl Select Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/acquisition/profiles/{profile_id}": {
      "get": {
        "tags": [
          "acquisition"
        ],
        "summary": "Api Acquisition Profile",
        "operationId": "api_acquisition_profile_api_v1_acquisition_profiles__profile_id__get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Api Acquisition Profile Api V1 Acquisition Profiles  Profile Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/acquisition/profiles/validate": {
      "post": {
        "tags": [
          "acquisition"
        ],
        "summary": "Api Validate Profile",
        "operationId": "api_validate_profile_api_v1_acquisition_profiles_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Profile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Api Validate Profile Api V1 Acquisition Profiles Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/acquisition/validation/start": {
      "post": {
        "tags": [
          "acquisition"
        ],
        "summary": "Api Validation Start",
        "operationId": "api_validation_start_api_v1_acquisition_validation_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Api Validation Start Api V1 Acquisition Validation Start Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/acquisition/validation/reports/{report_id}": {
      "get": {
        "tags": [
          "acquisition"
        ],
        "summary": "Api Validation Report",
        "operationId": "api_validation_report_api_v1_acquisition_validation_reports__report_id__get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Api Validation Report Api V1 Acquisition Validation Reports  Report Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcquisitionSelectRequest": {
        "properties": {
          "adapter": {
            "type": "string",
            "title": "Adapter",
            "description": "simulator, brainflow, lsl, csv_replay, or xdf_replay"
          },
          "config": {
            "type": "object",
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "adapter"
        ],
        "title": "AcquisitionSelectRequest"
      },
      "BrainFlowRequest": {
        "properties": {
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id",
            "default": "brainflow_synthetic_eeg"
          },
          "board_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Board Id"
          },
          "serial_port": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serial Port"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "mac_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mac Address"
          },
          "timeout": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout"
          },
          "start_stream": {
            "type": "boolean",
            "title": "Start Stream",
            "default": false
          }
        },
        "type": "object",
        "title": "BrainFlowRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LslSelectRequest": {
        "properties": {
          "stream_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Name"
          },
          "stream_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Type",
            "default": "EEG"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "timeout_seconds": {
            "type": "number",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Timeout Seconds",
            "default": 1.0
          }
        },
        "type": "object",
        "title": "LslSelectRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationStartRequest": {
        "properties": {
          "adapter": {
            "type": "string",
            "title": "Adapter",
            "default": "simulator"
          },
          "adapter_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adapter Type"
          },
          "stream_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Name"
          },
          "stream_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Type"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "marker_stream_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marker Stream Name"
          },
          "marker_stream_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marker Stream Type",
            "default": "Markers"
          },
          "board_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Board Id"
          },
          "serial_port": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serial Port"
          },
          "config": {
            "type": "object",
            "title": "Config"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "duration_seconds": {
            "type": "number",
            "maximum": 30.0,
            "minimum": 0.1,
            "title": "Duration Seconds",
            "default": 2.0
          },
          "record_windows": {
            "type": "boolean",
            "title": "Record Windows",
            "default": true
          },
          "run_sqi": {
            "type": "boolean",
            "title": "Run Sqi",
            "default": true
          },
          "run_shadow_inference": {
            "type": "boolean",
            "title": "Run Shadow Inference",
            "default": false
          }
        },
        "type": "object",
        "title": "ValidationStartRequest"
      }
    }
  }
}